wrap grpc server
